+++ e2e/e2etests/test_bitcoin_deposit.go
@@ -2,17 +2,26 @@ package e2etests
 
 import (
 	"fmt"
+	"strconv"
 
 	"github.com/zeta-chain/zetacore/e2e/runner"
 	"github.com/zeta-chain/zetacore/e2e/utils"
 	crosschaintypes "github.com/zeta-chain/zetacore/x/crosschain/types"
 )
 
-func TestBitcoinDeposit(r *runner.E2ERunner) {
+func TestBitcoinDeposit(r *runner.E2ERunner, args []string) {
+	if len(args) != 1 {
+		panic("TestBitcoinDeposit requires exactly one argument for the amount.")
+	}
+
+	depositAmount, err := strconv.ParseFloat(args[0], 64)
+	if err != nil {
+		panic("Invalid deposit amount specified for TestBitcoinDeposit.")
+	}
 
 	r.SetBtcAddress(r.Name, false)
 
-	txHash := r.DepositBTCWithAmount(0.001)
+	txHash := r.DepositBTCWithAmount(depositAmount)
 
 	// wait for the cctx to be mined
 	cctx := utils.WaitCctxMinedByInTxHash(r.Ctx, txHash.String(), r.CctxClient, r.Logger, r.CctxTimeout)