+++ e2e/e2etests/test_eth_withdraw.go
@@ -11,9 +11,23 @@ import (
 )
 
 // TestEtherWithdraw tests the withdraw of ether
-func TestEtherWithdraw(r *runner.E2ERunner) {
+func TestEtherWithdraw(r *runner.E2ERunner, args []string) {
+	approvedAmount := big.NewInt(1e18)
+	if len(args) != 1 {
+		panic("TestEtherWithdraw requires exactly one argument for the withdrawal amount.")
+	}
+
+	withdrawalAmount, ok := new(big.Int).SetString(args[0], 10)
+	if !ok {
+		panic("Invalid withdrawal amount specified for TestEtherWithdraw.")
+	}
+
+	if withdrawalAmount.Cmp(approvedAmount) >= 0 {
+		panic("Withdrawal amount must be less than the approved amount (1e18).")
+	}
+
 	// approve
-	tx, err := r.ETHZRC20.Approve(r.ZevmAuth, r.ETHZRC20Addr, big.NewInt(1e18))
+	tx, err := r.ETHZRC20.Approve(r.ZevmAuth, r.ETHZRC20Addr, approvedAmount)
 	if err != nil {
 		panic(err)
 	}
@@ -26,7 +40,7 @@ func TestEtherWithdraw(r *runner.E2ERunner) {
 	r.Logger.EVMReceipt(*receipt, "approve")
 
 	// withdraw
-	tx, err = r.ETHZRC20.Withdraw(r.ZevmAuth, r.DeployerAddress.Bytes(), big.NewInt(100000))
+	tx, err = r.ETHZRC20.Withdraw(r.ZevmAuth, r.DeployerAddress.Bytes(), withdrawalAmount)
 	if err != nil {
 		panic(err)
 	}
@@ -48,9 +62,23 @@ func TestEtherWithdraw(r *runner.E2ERunner) {
 }
 
 // TestEtherWithdrawRestricted tests the withdrawal to a restricted receiver address
-func TestEtherWithdrawRestricted(r *runner.E2ERunner) {
+func TestEtherWithdrawRestricted(r *runner.E2ERunner, args []string) {
+	approvedAmount := big.NewInt(1e18)
+	if len(args) != 1 {
+		panic("TestEtherWithdrawRestricted requires exactly one argument for the withdrawal amount.")
+	}
+
+	withdrawalAmount, ok := new(big.Int).SetString(args[0], 10)
+	if !ok {
+		panic("Invalid withdrawal amount specified for TestEtherWithdrawRestricted.")
+	}
+
+	if withdrawalAmount.Cmp(approvedAmount) >= 0 {
+		panic("Withdrawal amount must be less than the approved amount (1e18).")
+	}
+
 	// approve
-	tx, err := r.ETHZRC20.Approve(r.ZevmAuth, r.ETHZRC20Addr, big.NewInt(1e18))
+	tx, err := r.ETHZRC20.Approve(r.ZevmAuth, r.ETHZRC20Addr, approvedAmount)
 	if err != nil {
 		panic(err)
 	}
@@ -64,7 +92,7 @@ func TestEtherWithdrawRestricted(r *runner.E2ERunner) {
 
 	// withdraw
 	restrictedAddress := ethcommon.HexToAddress(testutils.RestrictedEVMAddressTest)
-	tx, err = r.ETHZRC20.Withdraw(r.ZevmAuth, restrictedAddress.Bytes(), big.NewInt(100000))
+	tx, err = r.ETHZRC20.Withdraw(r.ZevmAuth, restrictedAddress.Bytes(), withdrawalAmount)
 	if err != nil {
 		panic(err)
 	}