+++ typescript/authority/index.d.ts
@@ -1,0 +1,4 @@
+export * from "./genesis_pb";
+export * from "./policies_pb";
+export * from "./query_pb";
+export * from "./tx_pb";