+++ x/authority/client/cli/query_policies.go
@@ -1,0 +1,35 @@
+package cli
+
+import (
+	"context"
+
+	"github.com/cosmos/cosmos-sdk/client"
+	"github.com/cosmos/cosmos-sdk/client/flags"
+	"github.com/spf13/cobra"
+	"github.com/zeta-chain/zetacore/x/authority/types"
+)
+
+// CmdShowPolicies returns the command to show the policies
+func CmdShowPolicies() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "show-policies",
+		Short: "show the policies",
+		Args:  cobra.NoArgs,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx := client.GetClientContextFromCmd(cmd)
+
+			queryClient := types.NewQueryClient(clientCtx)
+
+			res, err := queryClient.Policies(context.Background(), &types.QueryGetPoliciesRequest{})
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(res)
+		},
+	}
+
+	flags.AddQueryFlagsToCmd(cmd)
+
+	return cmd
+}